requirements.txt
python-telegram-bot
openai
pypdf
moviepy
python-dotenv
requests
Pillow
google-generativeai
